+++ src/rungeKutta.h
@@ -23,26 +23,26 @@ Listplot[table[[;; , ;; 3 ;; 2]], ImageSize -> Large]
 //
 // =====================================
 
-std::vector<std::vector<float>> solveRungeKutta4th2eq
+std::vector<std::vector<double>> solveRungeKutta4th2eq
     (
-    float (*f1)(float, float, float),   // Equation for dy1/dt
-    float (*f2)(float, float, float),   // Equation for dy2/dt
+    double (*f1)(double, double, double),   // Equation for dy1/dt
+    double (*f2)(double, double, double),   // Equation for dy2/dt
     int n,                              // Number of steps
-    float t0, float tf,                 // Range of values of the independent variable
-    float y10, float y20                // Initial conditions
+    double t0, double tf,                 // Range of values of the independent variable
+    double y10, double y20                // Initial conditions
     );
 
-// std::vector<std::vector<float>> solveRungeKutta4th5eq
+// std::vector<std::vector<double>> solveRungeKutta4th5eq
 //     (
-//     float (*f1)(float, float, float),   // Equation for dy1/dt
-//     float (*f2)(float, float, float),   // Equation for dy2/dt
-//     float (*f3)(float, float, float),   // Equation for dy3/dt
-//     float (*f4)(float, float, float),   // Equation for dy4/dt
-//     float (*f5)(float, float, float),   // Equation for dy5/dt
+//     double (*f1)(double, double, double),   // Equation for dy1/dt
+//     double (*f2)(double, double, double),   // Equation for dy2/dt
+//     double (*f3)(double, double, double),   // Equation for dy3/dt
+//     double (*f4)(double, double, double),   // Equation for dy4/dt
+//     double (*f5)(double, double, double),   // Equation for dy5/dt
 //     int n,                              // Number of steps
-//     float lambda0, float lambdaf,       // Range of values of the independent variable
-//     float y10, float y20, float y30,    // Initial conditions
-//     float y40, float y50   
+//     double lambda0, double lambdaf,       // Range of values of the independent variable
+//     double y10, double y20, double y30,    // Initial conditions
+//     double y40, double y50   
 //     );
 
 //=====================================================
@@ -50,18 +50,18 @@ std::vector<std::vector<float>> solveRungeKutta4th2eq
 //  5 equations
 //=====================================================
 
-std::vector<std::vector<float>> solveRungeKutta4th5eqCustom
+std::vector<std::vector<double>> solveRungeKutta4th5eqCustom
     (
-    float (*f1)(float, float, float, float, float, float, float),   // Equation for dy1/dt
-    float (*f2)(float, float, float, float, float, float, float),   // Equation for dy2/dt
-    float (*f3)(float, float, float, float, float, float, float),   // Equation for dy3/dt
-    float (*f4)(float, float, float, float, float, float, float),   // Equation for dy4/dt
-    float (*f5)(float, float, float, float, float, float, float),   // Equation for dy5/dt
+    double (*f1)(double, double, double, double, double, double, double),   // Equation for dy1/dt
+    double (*f2)(double, double, double, double, double, double, double),   // Equation for dy2/dt
+    double (*f3)(double, double, double, double, double, double, double),   // Equation for dy3/dt
+    double (*f4)(double, double, double, double, double, double, double),   // Equation for dy4/dt
+    double (*f5)(double, double, double, double, double, double, double),   // Equation for dy5/dt
     int n,                              // Number of steps
-    float lambda0, float lambdaf,       // Range of values of the independent variable
-    float y10, float y20, float y30,    // Initial conditions
-    float y40, float y50,
-    float p_phi                         // Value of p_phi (constant)
+    double lambda0, double lambdaf,       // Range of values of the independent variable
+    double y10, double y20, double y30,    // Initial conditions
+    double y40, double y50,
+    double p_phi                         // Value of p_phi (constant)
     );
 
 #endif
